Move `get_joined_hosts` function into controller
